feat(diag): add chat suggestion feature
This commit adds a new feature to the diagnosis page that allows users to see suggested text replacements when hovering over highlighted text. The commit also includes changes to the HTML and CSS files, adding a comment box and styling it, as well as fixing some typos and updating the layout of the comments. Finally, a JavaScript file has been added to handle the suggestion feature.